refactor sgMail in utils
