+++ src/utils/index.js
@@ -1,3 +1,5 @@
 const { AppError, NotAuthorizedError, ConflictError, NotFoundError } = require("./appError");
+const {sendVerificationEmail, sendReVerificationEmail } = require("./mail");
 
-module.exports = { AppError, NotAuthorizedError, ConflictError, NotFoundError };
+module.exports = { AppError, NotAuthorizedError, ConflictError, NotFoundError, sendVerificationEmail, sendReVerificationEmail };
+
